[개선 - 500줄 초과 백엔드 모듈 분리]
video_utils, youtube_utils, task_service를 작은 서브모듈로 분리해 파일 크기 제한(500줄) 규칙을 충족시킴.

- video_utils.py (714 → 388줄): _video_encoder, _video_helpers, _video_transitions 추출
- youtube_utils.py (667 → 423줄): _youtube_parsers, _youtube_io, _youtube_downloader 추출
- services/task_service.py (562 → 482줄): _task_helpers, _task_notification_mixin 추출

기존 public API는 상위 모듈에서 re-export로 유지해 patch("src.*.*") 호환성 보존.
테스트 전 51개 통과.
