+++ backend/src/_video_helpers.py
@@ -1,0 +1,143 @@
+"""Sizing/timestamp/word helpers extracted from ``video_utils.py``.
+
+``video_utils`` re-exports each symbol so existing test patches
+(e.g. ``patch("src.video_utils.get_scaled_font_size")``) continue to work.
+"""
+
+from typing import Dict, List
+import logging
+
+logger = logging.getLogger(__name__)
+
+
+def format_ms_to_timestamp(ms: int) -> str:
+    """Format milliseconds to MM:SS format."""
+    seconds = ms // 1000
+    minutes = seconds // 60
+    seconds = seconds % 60
+    return f"{minutes:02d}:{seconds:02d}"
+
+
+def round_to_even(value: int) -> int:
+    """Round integer to nearest even number for H.264 compatibility."""
+    return value - (value % 2)
+
+
+def get_scaled_font_size(base_font_size: int, video_width: int) -> int:
+    """Scale caption font size by output width with sensible bounds."""
+    scaled_size = int(base_font_size * (video_width / 720))
+    return max(24, min(64, scaled_size))
+
+
+def get_subtitle_max_width(video_width: int) -> int:
+    """Return max subtitle text width with horizontal safe margins."""
+    horizontal_padding = max(40, int(video_width * 0.06))
+    return max(200, video_width - (horizontal_padding * 2))
+
+
+def get_safe_vertical_position(
+    video_height: int, text_height: int, position_y: float
+) -> int:
+    """Return subtitle y position clamped inside a top/bottom safe area."""
+    min_top_padding = max(40, int(video_height * 0.05))
+    min_bottom_padding = max(120, int(video_height * 0.10))
+
+    desired_y = int(video_height * position_y - text_height // 2)
+    max_y = video_height - min_bottom_padding - text_height
+    return max(min_top_padding, min(desired_y, max_y))
+
+
+def parse_timestamp_to_seconds(timestamp_str: str) -> float:
+    """Parse timestamp string to seconds."""
+    try:
+        timestamp_str = timestamp_str.strip()
+        logger.info(f"Parsing timestamp: '{timestamp_str}'")
+
+        if ":" in timestamp_str:
+            parts = timestamp_str.split(":")
+            if len(parts) == 2:
+                minutes, seconds = map(int, parts)
+                result = minutes * 60 + seconds
+                logger.info(f"Parsed '{timestamp_str}' -> {result}s")
+                return result
+            elif len(parts) == 3:
+                hours, minutes, seconds = map(int, parts)
+                result = hours * 3600 + minutes * 60 + seconds
+                logger.info(f"Parsed '{timestamp_str}' -> {result}s")
+                return result
+
+        result = float(timestamp_str)
+        logger.info(f"Parsed '{timestamp_str}' as seconds -> {result}s")
+        return result
+
+    except (ValueError, IndexError) as e:
+        logger.error(f"Failed to parse timestamp '{timestamp_str}': {e}")
+        return 0.0
+
+
+def get_words_in_range(
+    transcript_data: Dict, clip_start: float, clip_end: float
+) -> List[Dict]:
+    """Extract words that fall within a clip timerange."""
+    if not transcript_data or not transcript_data.get("words"):
+        return []
+
+    clip_start_ms = int(clip_start * 1000)
+    clip_end_ms = int(clip_end * 1000)
+
+    relevant_words = []
+    for word_data in transcript_data["words"]:
+        word_start = word_data["start"]
+        word_end = word_data["end"]
+
+        if word_start < clip_end_ms and word_end > clip_start_ms:
+            relative_start = max(0, (word_start - clip_start_ms) / 1000.0)
+            relative_end = min(
+                (clip_end_ms - clip_start_ms) / 1000.0,
+                (word_end - clip_start_ms) / 1000.0,
+            )
+
+            if relative_end > relative_start:
+                relevant_words.append(
+                    {
+                        "text": word_data["text"],
+                        "start": relative_start,
+                        "end": relative_end,
+                        "confidence": word_data.get("confidence", 1.0),
+                        "speaker": word_data.get("speaker"),
+                    }
+                )
+
+    return relevant_words
+
+
+def _chunk_words_by_speaker(
+    words: List[Dict], max_per_group: int
+) -> List[List[Dict]]:
+    """Group consecutive words, breaking at speaker changes or max group size.
+
+    Prevents two speakers from being rendered in a single subtitle line. When
+    diarization data is missing (speaker=None for all words) this behaves
+    identically to a fixed-size chunker.
+    """
+    if not words or max_per_group <= 0:
+        return []
+
+    groups: List[List[Dict]] = []
+    current: List[Dict] = []
+    current_speaker = words[0].get("speaker")
+
+    for word in words:
+        speaker = word.get("speaker")
+        if current and (speaker != current_speaker or len(current) >= max_per_group):
+            groups.append(current)
+            current = []
+            current_speaker = speaker
+        if not current:
+            current_speaker = speaker
+        current.append(word)
+
+    if current:
+        groups.append(current)
+
+    return groups
