+++ backend/src/services/_task_notification_mixin.py
@@ -1,0 +1,75 @@
+"""Task-completion email notification, extracted from ``TaskService``.
+
+The email symbols are resolved through the ``task_service`` module at call
+time so tests can continue to patch ``task_service.TaskCompletionEmailService``
+without seeing our import. ``self.db``, ``self.task_repo`` and ``self.config``
+come from ``TaskService.__init__``.
+"""
+
+from __future__ import annotations
+
+import logging
+
+logger = logging.getLogger(__name__)
+
+
+class _TaskNotificationMixin:
+    async def _send_completion_notification_if_needed(
+        self, *, task_id: str, clips_count: int
+    ) -> None:
+        from . import task_service as _ts
+
+        context = await self.task_repo.get_task_notification_context(self.db, task_id)
+        if not context:
+            logger.warning("Task %s missing notification context; skipping email", task_id)
+            return
+
+        if not context.get("notify_on_completion"):
+            return
+
+        if context.get("completion_notification_sent_at"):
+            logger.info(
+                "Completion notification already sent for task %s; skipping", task_id
+            )
+            return
+
+        user_email = context.get("user_email")
+        if not user_email:
+            logger.warning(
+                "Task %s has notify_on_completion enabled but user email is missing",
+                task_id,
+            )
+            return
+
+        email_service = _ts.TaskCompletionEmailService(self.config)
+        if not email_service.is_configured:
+            logger.warning(
+                "Skipping completion notification for task %s because Resend is not configured",
+                task_id,
+            )
+            return
+
+        try:
+            await email_service.send_task_completed_email(
+                recipient=_ts.TaskCompletionRecipient(
+                    email=user_email,
+                    name=context.get("user_name"),
+                    first_name=context.get("user_first_name"),
+                ),
+                task_id=task_id,
+                source_title=context.get("source_title"),
+                clips_count=clips_count,
+            )
+            stamped = await self.task_repo.mark_completion_notification_sent(
+                self.db, task_id
+            )
+            if not stamped:
+                logger.info(
+                    "Completion notification stamp already existed for task %s",
+                    task_id,
+                )
+        except Exception:
+            logger.exception(
+                "Failed to send completion notification for task %s",
+                task_id,
+            )
