+++ backend/src/services/task_service.py
@@ -8,7 +8,6 @@ import logging
 from datetime import datetime
 from pathlib import Path
 import json
-import hashlib
 from time import perf_counter
 
 import redis.asyncio as redis
@@ -17,25 +16,29 @@ from ..repositories.task_repository import TaskRepository
 from ..repositories.source_repository import SourceRepository
 from ..repositories.clip_repository import ClipRepository
 from ..repositories.cache_repository import CacheRepository
+from . import usage_service
+from ..usage_context import current_usage
 from .video_service import VideoService
-from .task_completion_email_service import (
+from .task_completion_email_service import (  # noqa: F401 — re-exported for tests to patch
     TaskCompletionEmailService,
     TaskCompletionRecipient,
 )
 from ..config import Config, get_config
-from ..clip_editor import (
-    trim_clip_file,
-    split_clip_file,
-    merge_clip_files,
-    overlay_custom_captions,
-)
-from ..video_utils import parse_timestamp_to_seconds
+from ._clip_edit_mixin import _ClipEditMixin
+from ._task_helpers import build_cache_key, is_queued_task_stale, seconds_to_mmss
+from ._task_notification_mixin import _TaskNotificationMixin
 
 logger = logging.getLogger(__name__)
 
 
-class TaskService:
-    """Service for task workflow orchestration."""
+class TaskService(_ClipEditMixin, _TaskNotificationMixin):
+    """Service for task workflow orchestration.
+
+    Clip-editing methods (``trim_clip``, ``split_clip``, ``merge_clips``,
+    ``update_clip_captions``, ``regenerate_all_clips_for_task``) live in
+    :class:`_ClipEditMixin`. Completion-email logic lives in
+    :class:`_TaskNotificationMixin`.
+    """
 
     def __init__(self, db: AsyncSession, config: Config | None = None):
         self.db = db
@@ -46,29 +49,10 @@ class TaskService:
         self.video_service = VideoService()
         self.config = config or get_config()
 
-    @staticmethod
-    def _build_cache_key(url: str, source_type: str, processing_mode: str) -> str:
-        payload = f"{source_type}|{processing_mode}|{url.strip()}"
-        return hashlib.sha256(payload.encode("utf-8")).hexdigest()
+    _build_cache_key = staticmethod(build_cache_key)
 
     def _is_stale_queued_task(self, task: Dict[str, Any]) -> bool:
-        """Detect queued tasks that have likely stalled due to worker issues."""
-        if task.get("status") != "queued":
-            return False
-
-        created_at = task.get("created_at")
-        updated_at = task.get("updated_at") or created_at
-
-        if not created_at or not updated_at:
-            return False
-
-        now = (
-            datetime.now(updated_at.tzinfo)
-            if getattr(updated_at, "tzinfo", None)
-            else datetime.utcnow()
-        )
-        age_seconds = (now - updated_at).total_seconds()
-        return age_seconds >= self.config.queued_task_timeout_seconds
+        return is_queued_task_stale(task, self.config.queued_task_timeout_seconds)
 
     async def create_task_with_source(
         self,
@@ -81,6 +65,11 @@ class TaskService:
         caption_template: str = "default",
         include_broll: bool = False,
         processing_mode: str = "fast",
+        stroke_color: Optional[str] = None,
+        stroke_width: Optional[int] = None,
+        subtitle_bold: bool = False,
+        subtitle_italic: bool = False,
+        subtitle_underline: bool = False,
     ) -> str:
         """
         Create a new task with associated source.
@@ -117,6 +106,11 @@ class TaskService:
             caption_template=caption_template,
             include_broll=include_broll,
             processing_mode=processing_mode,
+            stroke_color=stroke_color,
+            stroke_width=stroke_width,
+            subtitle_bold=subtitle_bold,
+            subtitle_italic=subtitle_italic,
+            subtitle_underline=subtitle_underline,
         )
 
         logger.info(f"Created task {task_id} for user {user_id}")
@@ -137,11 +131,21 @@ class TaskService:
         progress_callback: Optional[Callable] = None,
         should_cancel: Optional[Callable] = None,
         clip_ready_callback: Optional[Callable] = None,
+        stroke_color: Optional[str] = None,
+        stroke_width: Optional[int] = None,
+        bold: bool = False,
+        italic: bool = False,
+        underline: bool = False,
+        avoid_original_subtitle: str = "none",
     ) -> Dict[str, Any]:
         """
         Process a task: download video, analyze, create clips.
         Returns processing results.
         """
+        # Start recording API usage for this task. The ContextVar lives for
+        # the duration of this call; low-level code (AssemblyAI, LLM) reads
+        # it and appends events without knowing about task_service.
+        recorder, recorder_token = usage_service.start_recording()
         try:
             logger.info(f"Starting processing for task {task_id}")
             started_at = datetime.utcnow()
@@ -204,6 +208,11 @@ class TaskService:
                 cached_analysis_json=cached_analysis_json,
                 progress_callback=update_progress,
                 should_cancel=should_cancel,
+                stroke_color=stroke_color,
+                stroke_width=stroke_width,
+                bold=bold,
+                italic=italic,
+                underline=underline,
             )
             stage_timings["pipeline_seconds"] = round(
                 perf_counter() - pipeline_start, 3
@@ -254,6 +263,12 @@ class TaskService:
                     caption_template,
                     output_format,
                     add_subtitles,
+                    stroke_color=stroke_color,
+                    stroke_width=stroke_width,
+                    bold=bold,
+                    italic=italic,
+                    underline=underline,
+                    avoid_original_subtitle=avoid_original_subtitle,
                 )
                 if clip_info is None:
                     continue  # Skip failed clip
@@ -364,64 +379,20 @@ class TaskService:
                 error_code=error_code,
             )
             raise
-
-    async def _send_completion_notification_if_needed(
-        self, *, task_id: str, clips_count: int
-    ) -> None:
-        context = await self.task_repo.get_task_notification_context(self.db, task_id)
-        if not context:
-            logger.warning("Task %s missing notification context; skipping email", task_id)
-            return
-
-        if not context.get("notify_on_completion"):
-            return
-
-        if context.get("completion_notification_sent_at"):
-            logger.info(
-                "Completion notification already sent for task %s; skipping", task_id
-            )
-            return
-
-        user_email = context.get("user_email")
-        if not user_email:
-            logger.warning(
-                "Task %s has notify_on_completion enabled but user email is missing",
-                task_id,
-            )
-            return
-
-        email_service = TaskCompletionEmailService(self.config)
-        if not email_service.is_configured:
-            logger.warning(
-                "Skipping completion notification for task %s because Resend is not configured",
-                task_id,
-            )
-            return
-
-        try:
-            await email_service.send_task_completed_email(
-                recipient=TaskCompletionRecipient(
-                    email=user_email,
-                    name=context.get("user_name"),
-                    first_name=context.get("user_first_name"),
-                ),
-                task_id=task_id,
-                source_title=context.get("source_title"),
-                clips_count=clips_count,
-            )
-            stamped = await self.task_repo.mark_completion_notification_sent(
-                self.db, task_id
-            )
-            if not stamped:
-                logger.info(
-                    "Completion notification stamp already existed for task %s",
-                    task_id,
+        finally:
+            current_usage.reset(recorder_token)
+            try:
+                task_row = await self.task_repo.get_task_by_id(self.db, task_id)
+                task_user_id = task_row.get("user_id") if task_row else None
+                await usage_service.persist_recorder(
+                    self.db, recorder, task_id, task_user_id
                 )
-        except Exception:
-            logger.exception(
-                "Failed to send completion notification for task %s",
-                task_id,
-            )
+            except Exception as persist_exc:
+                logger.warning(
+                    "Failed to persist API usage for task %s: %s",
+                    task_id,
+                    persist_exc,
+                )
 
     async def get_task_with_clips(self, task_id: str) -> Optional[Dict[str, Any]]:
         """Get task details with all clips."""
@@ -504,294 +475,8 @@ class TaskService:
 
         return await self.get_task_with_clips(task_id) or {}
 
-    async def regenerate_all_clips_for_task(
-        self,
-        task_id: str,
-        font_family: str,
-        font_size: int,
-        font_color: str,
-        caption_template: str,
-    ) -> None:
-        """Regenerate all clips in a task using existing segment boundaries."""
-        task = await self.task_repo.get_task_by_id(self.db, task_id)
-        if not task:
-            raise ValueError("Task not found")
-
-        source_url = task.get("source_url")
-        source_type = task.get("source_type")
-        output_format = "vertical"
-        add_subtitles = True
-
-        # Preserve original output_format and add_subtitles from task creation (stored in Redis)
-        redis_client = redis.Redis(
-            host=self.config.redis_host,
-            port=self.config.redis_port,
-            password=self.config.redis_password,
-            decode_responses=True,
-        )
-        try:
-            source_payload = await redis_client.get(f"task_source:{task_id}")
-            if source_payload:
-                parsed = json.loads(source_payload)
-                of = parsed.get("output_format", output_format)
-                if of in ("vertical", "original"):
-                    output_format = of
-                asub = parsed.get("add_subtitles", add_subtitles)
-                if isinstance(asub, bool):
-                    add_subtitles = asub
-        finally:
-            await redis_client.close()
-
-        if not source_url or not source_type:
-            raise ValueError("Task source URL is missing; cannot regenerate clips")
-
-        clips = await self.clip_repo.get_clips_by_task(self.db, task_id)
-        if not clips:
-            return
-
-        video_path: Path
-        if source_type == "youtube":
-            downloaded = await self.video_service.download_video(source_url)
-            if not downloaded:
-                raise ValueError("Failed to download source video for regeneration")
-            video_path = Path(downloaded)
-        else:
-            video_path = self.video_service.resolve_local_video_path(source_url)
-            if not video_path.exists():
-                raise ValueError("Source video file no longer exists")
-
-        segments = [
-            {
-                "start_time": clip["start_time"],
-                "end_time": clip["end_time"],
-                "text": clip.get("text") or "",
-                "relevance_score": clip.get("relevance_score", 0.5),
-                "reasoning": clip.get("reasoning")
-                or "Regenerated with updated settings",
-                "virality_score": clip.get("virality_score", 0),
-                "hook_score": clip.get("hook_score", 0),
-                "engagement_score": clip.get("engagement_score", 0),
-                "value_score": clip.get("value_score", 0),
-                "shareability_score": clip.get("shareability_score", 0),
-                "hook_type": clip.get("hook_type"),
-            }
-            for clip in clips
-        ]
-
-        clips_info = await self.video_service.create_video_clips(
-            video_path,
-            segments,
-            font_family,
-            font_size,
-            font_color,
-            caption_template,
-            output_format,
-            add_subtitles,
-        )
-
-        await self.clip_repo.delete_clips_by_task(self.db, task_id)
-
-        clip_ids = []
-        for i, clip_info in enumerate(clips_info):
-            clip_id = await self.clip_repo.create_clip(
-                self.db,
-                task_id=task_id,
-                filename=clip_info["filename"],
-                file_path=clip_info["path"],
-                start_time=clip_info["start_time"],
-                end_time=clip_info["end_time"],
-                duration=clip_info["duration"],
-                text=clip_info.get("text") or "",
-                relevance_score=clip_info.get("relevance_score", 0.5),
-                reasoning=clip_info.get("reasoning")
-                or "Regenerated with updated settings",
-                clip_order=i + 1,
-                virality_score=clip_info.get("virality_score", 0),
-                hook_score=clip_info.get("hook_score", 0),
-                engagement_score=clip_info.get("engagement_score", 0),
-                value_score=clip_info.get("value_score", 0),
-                shareability_score=clip_info.get("shareability_score", 0),
-                hook_type=clip_info.get("hook_type"),
-            )
-            clip_ids.append(clip_id)
-
-        await self.task_repo.update_task_clips(self.db, task_id, clip_ids)
-
-    async def trim_clip(
-        self,
-        task_id: str,
-        clip_id: str,
-        start_offset: float,
-        end_offset: float,
-    ) -> Dict[str, Any]:
-        clip = await self.clip_repo.get_clip_by_id(self.db, clip_id)
-        if not clip or clip["task_id"] != task_id:
-            raise ValueError("Clip not found")
-
-        input_path = Path(clip["file_path"])
-        if not input_path.exists():
-            raise ValueError("Clip file not found")
-
-        output_path = trim_clip_file(
-            input_path, Path(self.config.temp_dir) / "clips", start_offset, end_offset
-        )
-        clip_duration = max(0.1, clip["duration"] - start_offset - end_offset)
-
-        start_seconds = parse_timestamp_to_seconds(clip["start_time"]) + start_offset
-        end_seconds = start_seconds + clip_duration
-
-        new_start = self._seconds_to_mmss(start_seconds)
-        new_end = self._seconds_to_mmss(end_seconds)
-
-        await self.clip_repo.update_clip(
-            self.db,
-            clip_id,
-            output_path.name,
-            str(output_path),
-            new_start,
-            new_end,
-            clip_duration,
-            clip.get("text") or "",
-        )
-        return (await self.clip_repo.get_clip_by_id(self.db, clip_id)) or {}
-
-    async def split_clip(
-        self, task_id: str, clip_id: str, split_time: float
-    ) -> Dict[str, Any]:
-        clip = await self.clip_repo.get_clip_by_id(self.db, clip_id)
-        if not clip or clip["task_id"] != task_id:
-            raise ValueError("Clip not found")
-
-        input_path = Path(clip["file_path"])
-        if not input_path.exists():
-            raise ValueError("Clip file not found")
-
-        first_path, second_path = split_clip_file(
-            input_path, Path(self.config.temp_dir) / "clips", split_time
-        )
-
-        start_seconds = parse_timestamp_to_seconds(clip["start_time"])
-        clamped_split = max(0.2, min(split_time, float(clip["duration"]) - 0.2))
-        split_abs = start_seconds + clamped_split
-        end_seconds = parse_timestamp_to_seconds(clip["end_time"])
-
-        await self.clip_repo.update_clip(
-            self.db,
-            clip_id,
-            first_path.name,
-            str(first_path),
-            clip["start_time"],
-            self._seconds_to_mmss(split_abs),
-            clamped_split,
-            clip.get("text") or "",
-        )
-
-        await self.clip_repo.create_clip(
-            self.db,
-            task_id=task_id,
-            filename=second_path.name,
-            file_path=str(second_path),
-            start_time=self._seconds_to_mmss(split_abs),
-            end_time=self._seconds_to_mmss(end_seconds),
-            duration=max(0.1, end_seconds - split_abs),
-            text=clip.get("text") or "",
-            relevance_score=clip.get("relevance_score", 0.5),
-            reasoning=clip.get("reasoning") or "Split from original clip",
-            clip_order=clip.get("clip_order", 1) + 1,
-            virality_score=clip.get("virality_score", 0),
-            hook_score=clip.get("hook_score", 0),
-            engagement_score=clip.get("engagement_score", 0),
-            value_score=clip.get("value_score", 0),
-            shareability_score=clip.get("shareability_score", 0),
-            hook_type=clip.get("hook_type"),
-        )
-
-        await self.clip_repo.reorder_task_clips(self.db, task_id)
-        return {"message": "Clip split successfully"}
-
-    async def merge_clips(self, task_id: str, clip_ids: list[str]) -> Dict[str, Any]:
-        if len(clip_ids) < 2:
-            raise ValueError("At least two clips are required to merge")
-
-        clips = []
-        for clip_id in clip_ids:
-            clip = await self.clip_repo.get_clip_by_id(self.db, clip_id)
-            if not clip or clip["task_id"] != task_id:
-                raise ValueError("One or more clips not found")
-            clips.append(clip)
-
-        ordered = sorted(clips, key=lambda c: c.get("clip_order", 0))
-        merged_path = merge_clip_files(
-            [Path(c["file_path"]) for c in ordered],
-            Path(self.config.temp_dir) / "clips",
-        )
-
-        start_time = ordered[0]["start_time"]
-        end_time = ordered[-1]["end_time"]
-        duration = sum(float(c.get("duration", 0.0)) for c in ordered)
-        text = " ".join((c.get("text") or "").strip() for c in ordered if c.get("text"))
-
-        first = ordered[0]
-        await self.clip_repo.update_clip(
-            self.db,
-            first["id"],
-            merged_path.name,
-            str(merged_path),
-            start_time,
-            end_time,
-            duration,
-            text,
-        )
-
-        for clip in ordered[1:]:
-            await self.clip_repo.delete_clip(self.db, clip["id"])
-
-        await self.clip_repo.reorder_task_clips(self.db, task_id)
-        return {"message": "Clips merged successfully", "clip_id": first["id"]}
-
-    async def update_clip_captions(
-        self,
-        task_id: str,
-        clip_id: str,
-        caption_text: str,
-        position: str,
-        highlight_words: list[str],
-    ) -> Dict[str, Any]:
-        clip = await self.clip_repo.get_clip_by_id(self.db, clip_id)
-        if not clip or clip["task_id"] != task_id:
-            raise ValueError("Clip not found")
-
-        input_path = Path(clip["file_path"])
-        if not input_path.exists():
-            raise ValueError("Clip file not found")
-
-        output_path = overlay_custom_captions(
-            input_path,
-            Path(self.config.temp_dir) / "clips",
-            caption_text,
-            position,
-            highlight_words,
-        )
-
-        await self.clip_repo.update_clip(
-            self.db,
-            clip_id,
-            output_path.name,
-            str(output_path),
-            clip["start_time"],
-            clip["end_time"],
-            clip["duration"],
-            caption_text,
-        )
-        return (await self.clip_repo.get_clip_by_id(self.db, clip_id)) or {}
-
     async def get_performance_metrics(self) -> Dict[str, Any]:
         """Return aggregate processing performance metrics."""
         return await self.task_repo.get_performance_metrics(self.db)
 
-    @staticmethod
-    def _seconds_to_mmss(seconds: float) -> str:
-        total = max(0, int(round(seconds)))
-        minutes = total // 60
-        secs = total % 60
-        return f"{minutes:02d}:{secs:02d}"
+    _seconds_to_mmss = staticmethod(seconds_to_mmss)
