+++ backend/src/video_utils.py
@@ -1,1153 +1,82 @@
 """
 Utility functions for video-related operations.
-Optimized for MoviePy v2, AssemblyAI integration, and high-quality output.
+
+Most of the heavy lifting lives in sibling ``_video_*`` modules; this file now
+hosts the public entry points (``create_optimized_clip``,
+``create_clips_from_segments``) and re-exports the helpers so existing
+``patch("src.video_utils.<symbol>")`` tests keep working.
 """
 
 from pathlib import Path
-from typing import List, Dict, Any, Tuple, Optional
-import os
+from typing import Any, Dict, List, Optional
 import logging
-import numpy as np
-from concurrent.futures import ThreadPoolExecutor
-import json
 
-import cv2
-from moviepy import VideoFileClip, CompositeVideoClip, TextClip, ColorClip
+from moviepy import (
+    ColorClip,
+    CompositeVideoClip,
+    TextClip,
+    VideoFileClip,
+    concatenate_videoclips,
+)
 from moviepy.video.fx import CrossFadeIn, CrossFadeOut, FadeIn, FadeOut
 
 import assemblyai as aai
-import srt
-from datetime import timedelta
+import srt  # noqa: F401 — re-exported for backwards-compatible patching
+from datetime import timedelta  # noqa: F401
 
 from .config import Config
-from .caption_templates import get_template, CAPTION_TEMPLATES
-from .font_registry import find_font_path
+from .caption_templates import get_template, CAPTION_TEMPLATES  # noqa: F401
+from .font_registry import find_font_path  # noqa: F401
+from ._video_encoder import VideoProcessor, _resolve_font_with_style  # noqa: F401
+from ._video_helpers import (
+    _chunk_words_by_speaker,  # noqa: F401
+    format_ms_to_timestamp,  # noqa: F401
+    get_safe_vertical_position,  # noqa: F401
+    get_scaled_font_size,  # noqa: F401
+    get_subtitle_max_width,  # noqa: F401
+    get_words_in_range,  # noqa: F401
+    parse_timestamp_to_seconds,
+    round_to_even,
+)
+from ._video_transitions import (  # noqa: F401
+    apply_transition_effect,
+    get_available_transitions,
+)
+from ._video_face import (
+    detect_faces_in_clip,  # noqa: F401
+    detect_optimal_crop_region,
+    filter_face_outliers,  # noqa: F401
+)
+from ._video_broll import (  # noqa: F401
+    apply_broll_to_clip,
+    create_9_16_clip,
+    insert_broll_into_clip,
+    resize_for_916,
+)
+from ._video_subtitles import (  # noqa: F401
+    create_assemblyai_subtitles,
+    create_fade_subtitles,
+    create_karaoke_subtitles,
+    create_pop_subtitles,
+    create_static_subtitles,
+)
+from ._video_transcript import (
+    cache_transcript_data,  # noqa: F401
+    format_transcript_for_analysis,  # noqa: F401
+    get_video_transcript,
+    load_cached_transcript_data,  # noqa: F401
+)
 
 logger = logging.getLogger(__name__)
 config = Config()
 TRANSCRIPT_CACHE_SCHEMA_VERSION = 2
 
 
-class VideoProcessor:
-    """Handles video processing operations with optimized settings."""
-
-    def __init__(
-        self,
-        font_family: str = "THEBOLDFONT",
-        font_size: int = 24,
-        font_color: str = "#FFFFFF",
-    ):
-        self.font_family = font_family
-        self.font_size = font_size
-        self.font_color = font_color
-        resolved_font = find_font_path(font_family, allow_all_user_fonts=True)
-        if not resolved_font:
-            resolved_font = find_font_path("TikTokSans-Regular")
-        if not resolved_font:
-            resolved_font = find_font_path("THEBOLDFONT")
-        self.font_path = str(resolved_font) if resolved_font else ""
-
-    def get_optimal_encoding_settings(
-        self, target_quality: str = "high"
-    ) -> Dict[str, Any]:
-        """Get optimal encoding settings for different quality levels."""
-        settings = {
-            "high": {
-                "codec": "libx264",
-                "audio_codec": "aac",
-                "audio_bitrate": "256k",
-                "preset": "slow",
-                "ffmpeg_params": [
-                    "-crf",
-                    "18",
-                    "-pix_fmt",
-                    "yuv420p",
-                    "-profile:v",
-                    "high",
-                    "-movflags",
-                    "+faststart",
-                    "-sws_flags",
-                    "lanczos",
-                ],
-            },
-            "medium": {
-                "codec": "libx264",
-                "audio_codec": "aac",
-                "bitrate": "4000k",
-                "audio_bitrate": "192k",
-                "preset": "fast",
-                "ffmpeg_params": ["-crf", "23", "-pix_fmt", "yuv420p"],
-            },
-        }
-        return settings.get(target_quality, settings["high"])
-
-
-def get_video_transcript(video_path: Path, speech_model: str = "best") -> str:
-    """Get transcript using AssemblyAI with word-level timing for precise subtitles."""
-    logger.info(f"Getting transcript for: {video_path}")
-
-    # Configure AssemblyAI
-    aai.settings.api_key = config.assembly_ai_api_key
-    transcriber = aai.Transcriber()
-
-    # Request word-level timestamps for precise subtitle sync
-    speech_model_value = aai.SpeechModel.best
-    if speech_model == "nano":
-        speech_model_value = aai.SpeechModel.nano
-
-    config_obj = aai.TranscriptionConfig(
-        speaker_labels=True,
-        punctuate=True,
-        format_text=True,
-        speech_model=speech_model_value,
-    )
-
-    try:
-        logger.info("Starting AssemblyAI transcription")
-        transcript = transcriber.transcribe(str(video_path), config=config_obj)
-
-        if transcript.status == aai.TranscriptStatus.error:
-            logger.error(f"AssemblyAI transcription failed: {transcript.error}")
-            raise Exception(f"Transcription failed: {transcript.error}")
-
-        formatted_lines = format_transcript_for_analysis(transcript)
-
-        # Cache the raw transcript for subtitle generation
-        cache_transcript_data(video_path, transcript)
-
-        result = "\n".join(formatted_lines)
-        logger.info(
-            f"Transcript formatted: {len(formatted_lines)} segments, {len(result)} chars"
-        )
-        return result
-
-    except Exception as e:
-        logger.error(f"Error in transcription: {e}")
-        raise
-
-
-def cache_transcript_data(video_path: Path, transcript) -> None:
-    """Cache AssemblyAI transcript data for subtitle generation."""
-    cache_path = video_path.with_suffix(".transcript_cache.json")
-
-    words_data = []
-    if transcript.words:
-        words_data = [_serialize_transcript_word(word) for word in transcript.words]
-
-    utterances_data = []
-    if getattr(transcript, "utterances", None):
-        utterances_data = [
-            {
-                "text": utterance.text,
-                "start": utterance.start,
-                "end": utterance.end,
-                "speaker": getattr(utterance, "speaker", None),
-                "words": [
-                    _serialize_transcript_word(word)
-                    for word in getattr(utterance, "words", []) or []
-                ],
-            }
-            for utterance in transcript.utterances
-        ]
-
-    cache_data = {
-        "version": TRANSCRIPT_CACHE_SCHEMA_VERSION,
-        "words": words_data,
-        "utterances": utterances_data,
-        "text": transcript.text,
-    }
-
-    with open(cache_path, "w") as f:
-        json.dump(cache_data, f)
-
-    logger.info(f"Cached {len(words_data)} words to {cache_path}")
-
-
-def load_cached_transcript_data(video_path: Path) -> Optional[Dict]:
-    """Load cached AssemblyAI transcript data."""
-    cache_path = video_path.with_suffix(".transcript_cache.json")
-
-    if not cache_path.exists():
-        return None
-
-    try:
-        with open(cache_path, "r") as f:
-            payload = json.load(f)
-            if "version" not in payload:
-                payload["version"] = TRANSCRIPT_CACHE_SCHEMA_VERSION
-                payload.setdefault("utterances", [])
-            return payload
-    except Exception as e:
-        logger.warning(f"Failed to load transcript cache: {e}")
-        return None
-
-
-def _serialize_transcript_word(word) -> Dict[str, Any]:
-    return {
-        "text": word.text,
-        "start": word.start,
-        "end": word.end,
-        "confidence": word.confidence if hasattr(word, "confidence") else 1.0,
-        "speaker": getattr(word, "speaker", None),
-    }
-
-
-def format_transcript_for_analysis(transcript) -> List[str]:
-    """Format transcripts into readable timestamped segments for AI analysis."""
-    utterances = getattr(transcript, "utterances", None) or []
-    if utterances:
-        formatted_lines = []
-        for utterance in utterances:
-            start_time = format_ms_to_timestamp(utterance.start)
-            end_time = format_ms_to_timestamp(utterance.end)
-            speaker = getattr(utterance, "speaker", None)
-            speaker_prefix = f"Speaker {speaker}: " if speaker else ""
-            formatted_lines.append(
-                f"[{start_time} - {end_time}] {speaker_prefix}{utterance.text}"
-            )
-        return formatted_lines
-
-    formatted_lines = []
-    words = getattr(transcript, "words", None) or []
-    if not words:
-        return formatted_lines
-
-    logger.info(f"Processing {len(words)} words with precise timing")
-
-    current_segment = []
-    current_start = None
-    segment_word_count = 0
-    max_words_per_segment = 8
-
-    for word in words:
-        if current_start is None:
-            current_start = word.start
-
-        current_segment.append(word.text)
-        segment_word_count += 1
-
-        if (
-            segment_word_count >= max_words_per_segment
-            or word.text.endswith(".")
-            or word.text.endswith("!")
-            or word.text.endswith("?")
-        ):
-            if current_segment:
-                start_time = format_ms_to_timestamp(current_start)
-                end_time = format_ms_to_timestamp(word.end)
-                text = " ".join(current_segment)
-                formatted_lines.append(f"[{start_time} - {end_time}] {text}")
-
-            current_segment = []
-            current_start = None
-            segment_word_count = 0
-
-    if current_segment and current_start is not None:
-        start_time = format_ms_to_timestamp(current_start)
-        end_time = format_ms_to_timestamp(words[-1].end)
-        text = " ".join(current_segment)
-        formatted_lines.append(f"[{start_time} - {end_time}] {text}")
-
-    return formatted_lines
-
-
-def format_ms_to_timestamp(ms: int) -> str:
-    """Format milliseconds to MM:SS format."""
-    seconds = ms // 1000
-    minutes = seconds // 60
-    seconds = seconds % 60
-    return f"{minutes:02d}:{seconds:02d}"
-
-
-def round_to_even(value: int) -> int:
-    """Round integer to nearest even number for H.264 compatibility."""
-    return value - (value % 2)
-
-
-def get_scaled_font_size(base_font_size: int, video_width: int) -> int:
-    """Scale caption font size by output width with sensible bounds."""
-    scaled_size = int(base_font_size * (video_width / 720))
-    return max(24, min(64, scaled_size))
-
-
-def get_subtitle_max_width(video_width: int) -> int:
-    """Return max subtitle text width with horizontal safe margins."""
-    horizontal_padding = max(40, int(video_width * 0.06))
-    return max(200, video_width - (horizontal_padding * 2))
-
-
-def get_safe_vertical_position(
-    video_height: int, text_height: int, position_y: float
-) -> int:
-    """Return subtitle y position clamped inside a top/bottom safe area."""
-    min_top_padding = max(40, int(video_height * 0.05))
-    min_bottom_padding = max(120, int(video_height * 0.10))
-
-    desired_y = int(video_height * position_y - text_height // 2)
-    max_y = video_height - min_bottom_padding - text_height
-    return max(min_top_padding, min(desired_y, max_y))
-
-
-def detect_optimal_crop_region(
-    video_clip: VideoFileClip,
-    start_time: float,
-    end_time: float,
-    target_ratio: float = 9 / 16,
-) -> Tuple[int, int, int, int]:
-    """Detect optimal crop region using improved face detection."""
-    try:
-        original_width, original_height = video_clip.size
-
-        # Calculate target dimensions and ensure they're even
-        if original_width / original_height > target_ratio:
-            new_width = round_to_even(int(original_height * target_ratio))
-            new_height = round_to_even(original_height)
-        else:
-            new_width = round_to_even(original_width)
-            new_height = round_to_even(int(original_width / target_ratio))
-
-        # Try improved face detection
-        face_centers = detect_faces_in_clip(video_clip, start_time, end_time)
-
-        # Calculate crop position
-        if face_centers:
-            # Use weighted average of face centers with temporal consistency
-            total_weight = sum(
-                area * confidence for _, _, area, confidence in face_centers
-            )
-            if total_weight > 0:
-                weighted_x = (
-                    sum(
-                        x * area * confidence for x, y, area, confidence in face_centers
-                    )
-                    / total_weight
-                )
-                weighted_y = (
-                    sum(
-                        y * area * confidence for x, y, area, confidence in face_centers
-                    )
-                    / total_weight
-                )
-
-                # Add slight bias towards upper portion for better face framing
-                weighted_y = max(0, weighted_y - new_height * 0.1)
-
-                x_offset = max(
-                    0, min(int(weighted_x - new_width // 2), original_width - new_width)
-                )
-                y_offset = max(
-                    0,
-                    min(
-                        int(weighted_y - new_height // 2), original_height - new_height
-                    ),
-                )
-
-                logger.info(
-                    f"Face-centered crop: {len(face_centers)} faces detected with improved algorithm"
-                )
-            else:
-                # Center crop
-                x_offset = (
-                    (original_width - new_width) // 2
-                    if original_width > new_width
-                    else 0
-                )
-                y_offset = (
-                    (original_height - new_height) // 2
-                    if original_height > new_height
-                    else 0
-                )
-        else:
-            # Center crop
-            x_offset = (
-                (original_width - new_width) // 2 if original_width > new_width else 0
-            )
-            y_offset = (
-                (original_height - new_height) // 2
-                if original_height > new_height
-                else 0
-            )
-            logger.info("Using center crop (no faces detected)")
-
-        # Ensure offsets are even too
-        x_offset = round_to_even(x_offset)
-        y_offset = round_to_even(y_offset)
-
-        logger.info(
-            f"Crop dimensions: {new_width}x{new_height} at offset ({x_offset}, {y_offset})"
-        )
-        return (x_offset, y_offset, new_width, new_height)
-
-    except Exception as e:
-        logger.error(f"Error in crop detection: {e}")
-        # Fallback to center crop
-        original_width, original_height = video_clip.size
-        if original_width / original_height > target_ratio:
-            new_width = round_to_even(int(original_height * target_ratio))
-            new_height = round_to_even(original_height)
-        else:
-            new_width = round_to_even(original_width)
-            new_height = round_to_even(int(original_width / target_ratio))
-
-        x_offset = (
-            round_to_even((original_width - new_width) // 2)
-            if original_width > new_width
-            else 0
-        )
-        y_offset = (
-            round_to_even((original_height - new_height) // 2)
-            if original_height > new_height
-            else 0
-        )
-
-        return (x_offset, y_offset, new_width, new_height)
-
-
-def detect_faces_in_clip(
-    video_clip: VideoFileClip, start_time: float, end_time: float
-) -> List[Tuple[int, int, int, float]]:
-    """
-    Improved face detection using multiple methods and temporal consistency.
-    Returns list of (x, y, area, confidence) tuples.
-    """
-    face_centers = []
-
-    try:
-        # Try to use MediaPipe (most accurate)
-        mp_face_detection = None
-        try:
-            import mediapipe as mp
-
-            mp_face_detection = mp.solutions.face_detection.FaceDetection(
-                model_selection=0,  # 0 for short-range (better for close faces)
-                min_detection_confidence=0.5,
-            )
-            logger.info("Using MediaPipe face detector")
-        except ImportError:
-            logger.info("MediaPipe not available, falling back to OpenCV")
-        except Exception as e:
-            logger.warning(f"MediaPipe face detector failed to initialize: {e}")
-
-        # Initialize OpenCV face detectors as fallback
-        haar_cascade = cv2.CascadeClassifier(
-            cv2.data.haarcascades + "haarcascade_frontalface_default.xml"
-        )
-
-        # Try to load DNN face detector (more accurate than Haar)
-        dnn_net = None
-        try:
-            # Load OpenCV's DNN face detector
-            prototxt_path = cv2.data.haarcascades.replace(
-                "haarcascades", "opencv_face_detector.pbtxt"
-            )
-            model_path = cv2.data.haarcascades.replace(
-                "haarcascades", "opencv_face_detector_uint8.pb"
-            )
-
-            # If DNN model files don't exist, we'll fall back to Haar cascade
-            import os
-
-            if os.path.exists(prototxt_path) and os.path.exists(model_path):
-                dnn_net = cv2.dnn.readNetFromTensorflow(model_path, prototxt_path)
-                logger.info("OpenCV DNN face detector loaded as backup")
-            else:
-                logger.info("OpenCV DNN face detector not available")
-        except Exception:
-            logger.info("OpenCV DNN face detector failed to load")
-
-        # Sample more frames for better face detection (every 0.5 seconds)
-        duration = end_time - start_time
-        sample_interval = min(0.5, duration / 10)  # At least 10 samples, max every 0.5s
-        sample_times = []
-
-        current_time = start_time
-        while current_time < end_time:
-            sample_times.append(current_time)
-            current_time += sample_interval
-
-        # Ensure we always sample the middle and end
-        if duration > 1.0:
-            middle_time = start_time + duration / 2
-            if middle_time not in sample_times:
-                sample_times.append(middle_time)
-
-        sample_times = [t for t in sample_times if t < end_time]
-        logger.info(f"Sampling {len(sample_times)} frames for face detection")
-
-        for sample_time in sample_times:
-            try:
-                frame = video_clip.get_frame(sample_time)
-                height, width = frame.shape[:2]
-                detected_faces = []
-
-                # Try MediaPipe first (most accurate)
-                if mp_face_detection is not None:
-                    try:
-                        # MediaPipe expects RGB format
-                        results = mp_face_detection.process(frame)
-
-                        if results.detections:
-                            for detection in results.detections:
-                                bbox = detection.location_data.relative_bounding_box
-                                confidence = detection.score[0]
-
-                                # Convert relative coordinates to absolute
-                                x = int(bbox.xmin * width)
-                                y = int(bbox.ymin * height)
-                                w = int(bbox.width * width)
-                                h = int(bbox.height * height)
-
-                                if w > 30 and h > 30:  # Minimum face size
-                                    detected_faces.append((x, y, w, h, confidence))
-                    except Exception as e:
-                        logger.warning(
-                            f"MediaPipe detection failed for frame at {sample_time}s: {e}"
-                        )
-
-                # If MediaPipe didn't find faces, try DNN detector
-                if not detected_faces and dnn_net is not None:
-                    try:
-                        frame_bgr = cv2.cvtColor(frame, cv2.COLOR_RGB2BGR)
-                        blob = cv2.dnn.blobFromImage(
-                            frame_bgr, 1.0, (300, 300), [104, 117, 123]
-                        )
-                        dnn_net.setInput(blob)
-                        detections = dnn_net.forward()
-
-                        for i in range(detections.shape[2]):
-                            confidence = detections[0, 0, i, 2]
-                            if confidence > 0.5:  # Confidence threshold
-                                x1 = int(detections[0, 0, i, 3] * width)
-                                y1 = int(detections[0, 0, i, 4] * height)
-                                x2 = int(detections[0, 0, i, 5] * width)
-                                y2 = int(detections[0, 0, i, 6] * height)
-
-                                w = x2 - x1
-                                h = y2 - y1
-
-                                if w > 30 and h > 30:  # Minimum face size
-                                    detected_faces.append((x1, y1, w, h, confidence))
-                    except Exception as e:
-                        logger.warning(
-                            f"DNN detection failed for frame at {sample_time}s: {e}"
-                        )
-
-                # If still no faces found, use Haar cascade
-                if not detected_faces:
-                    try:
-                        frame_bgr = cv2.cvtColor(frame, cv2.COLOR_RGB2BGR)
-                        gray = cv2.cvtColor(frame_bgr, cv2.COLOR_BGR2GRAY)
-
-                        faces = haar_cascade.detectMultiScale(
-                            gray,
-                            scaleFactor=1.05,  # More sensitive
-                            minNeighbors=3,  # Less strict
-                            minSize=(40, 40),  # Smaller minimum size
-                            maxSize=(
-                                int(width * 0.7),
-                                int(height * 0.7),
-                            ),  # Maximum size limit
-                        )
-
-                        for x, y, w, h in faces:
-                            # Estimate confidence based on face size and position
-                            face_area = w * h
-                            relative_size = face_area / (width * height)
-                            confidence = min(
-                                0.9, 0.3 + relative_size * 2
-                            )  # Rough confidence estimate
-                            detected_faces.append((x, y, w, h, confidence))
-                    except Exception as e:
-                        logger.warning(
-                            f"Haar cascade detection failed for frame at {sample_time}s: {e}"
-                        )
-
-                # Process detected faces
-                for x, y, w, h, confidence in detected_faces:
-                    face_center_x = x + w // 2
-                    face_center_y = y + h // 2
-                    face_area = w * h
-
-                    # Filter out very small or very large faces
-                    frame_area = width * height
-                    relative_area = face_area / frame_area
-
-                    if (
-                        0.005 < relative_area < 0.3
-                    ):  # Face should be 0.5% to 30% of frame
-                        face_centers.append(
-                            (face_center_x, face_center_y, face_area, confidence)
-                        )
-
-            except Exception as e:
-                logger.warning(f"Error detecting faces in frame at {sample_time}s: {e}")
-                continue
-
-        # Close MediaPipe detector
-        if mp_face_detection is not None:
-            mp_face_detection.close()
-
-        # Remove outliers (faces that are very far from the median position)
-        if len(face_centers) > 2:
-            face_centers = filter_face_outliers(face_centers)
-
-        logger.info(f"Detected {len(face_centers)} reliable face centers")
-        return face_centers
-
-    except Exception as e:
-        logger.error(f"Error in face detection: {e}")
-        return []
-
-
-def filter_face_outliers(
-    face_centers: List[Tuple[int, int, int, float]],
-) -> List[Tuple[int, int, int, float]]:
-    """Remove face detections that are outliers (likely false positives)."""
-    if len(face_centers) < 3:
-        return face_centers
-
-    try:
-        # Calculate median position
-        x_positions = [x for x, y, area, conf in face_centers]
-        y_positions = [y for x, y, area, conf in face_centers]
-
-        median_x = np.median(x_positions)
-        median_y = np.median(y_positions)
-
-        # Calculate standard deviation
-        std_x = np.std(x_positions)
-        std_y = np.std(y_positions)
-
-        # Filter out faces that are more than 2 standard deviations away
-        filtered_faces = []
-        for face in face_centers:
-            x, y, area, conf = face
-            if abs(x - median_x) <= 2 * std_x and abs(y - median_y) <= 2 * std_y:
-                filtered_faces.append(face)
-
-        logger.info(
-            f"Filtered {len(face_centers)} -> {len(filtered_faces)} faces (removed outliers)"
-        )
-        return (
-            filtered_faces if filtered_faces else face_centers
-        )  # Return original if all filtered
-
-    except Exception as e:
-        logger.warning(f"Error filtering face outliers: {e}")
-        return face_centers
-
-
-def parse_timestamp_to_seconds(timestamp_str: str) -> float:
-    """Parse timestamp string to seconds."""
-    try:
-        timestamp_str = timestamp_str.strip()
-        logger.info(f"Parsing timestamp: '{timestamp_str}'")  # Debug logging
-
-        if ":" in timestamp_str:
-            parts = timestamp_str.split(":")
-            if len(parts) == 2:
-                minutes, seconds = map(int, parts)
-                result = minutes * 60 + seconds
-                logger.info(f"Parsed '{timestamp_str}' -> {result}s")
-                return result
-            elif len(parts) == 3:  # HH:MM:SS format
-                hours, minutes, seconds = map(int, parts)
-                result = hours * 3600 + minutes * 60 + seconds
-                logger.info(f"Parsed '{timestamp_str}' -> {result}s")
-                return result
-
-        # Try parsing as pure seconds
-        result = float(timestamp_str)
-        logger.info(f"Parsed '{timestamp_str}' as seconds -> {result}s")
-        return result
-
-    except (ValueError, IndexError) as e:
-        logger.error(f"Failed to parse timestamp '{timestamp_str}': {e}")
-        return 0.0
-
-
-def get_words_in_range(
-    transcript_data: Dict, clip_start: float, clip_end: float
-) -> List[Dict]:
-    """Extract words that fall within a clip timerange."""
-    if not transcript_data or not transcript_data.get("words"):
-        return []
-
-    clip_start_ms = int(clip_start * 1000)
-    clip_end_ms = int(clip_end * 1000)
-
-    relevant_words = []
-    for word_data in transcript_data["words"]:
-        word_start = word_data["start"]
-        word_end = word_data["end"]
-
-        if word_start < clip_end_ms and word_end > clip_start_ms:
-            relative_start = max(0, (word_start - clip_start_ms) / 1000.0)
-            relative_end = min(
-                (clip_end_ms - clip_start_ms) / 1000.0,
-                (word_end - clip_start_ms) / 1000.0,
-            )
-
-            if relative_end > relative_start:
-                relevant_words.append(
-                    {
-                        "text": word_data["text"],
-                        "start": relative_start,
-                        "end": relative_end,
-                        "confidence": word_data.get("confidence", 1.0),
-                    }
-                )
-
-    return relevant_words
-
-
-def create_assemblyai_subtitles(
-    video_path: Path,
-    clip_start: float,
-    clip_end: float,
-    video_width: int,
-    video_height: int,
-    font_family: str = "THEBOLDFONT",
-    font_size: int = 24,
-    font_color: str = "#FFFFFF",
-    caption_template: str = "default",
-) -> List[TextClip]:
-    """Create subtitles using AssemblyAI's precise word timing with template support."""
-    transcript_data = load_cached_transcript_data(video_path)
-
-    if not transcript_data or not transcript_data.get("words"):
-        logger.warning("No cached transcript data available for subtitles")
-        return []
-
-    # Get template settings
-    template = get_template(caption_template)
-    animation_type = template.get("animation", "none")
-
-    effective_font_family = font_family or template["font_family"]
-    effective_font_size = int(font_size) if font_size else int(template["font_size"])
-    effective_font_color = font_color or template["font_color"]
-    effective_template = {
-        **template,
-        "font_size": effective_font_size,
-        "font_color": effective_font_color,
-        "font_family": effective_font_family,
-    }
-
-    logger.info(
-        f"Creating subtitles with template '{caption_template}', animation: {animation_type}"
-    )
-
-    # Get words in range
-    relevant_words = get_words_in_range(transcript_data, clip_start, clip_end)
-
-    if not relevant_words:
-        logger.warning("No words found in clip timerange")
-        return []
-
-    # Choose subtitle creation method based on animation type
-    if animation_type == "karaoke":
-        return create_karaoke_subtitles(
-            relevant_words,
-            video_width,
-            video_height,
-            effective_template,
-            effective_font_family,
-        )
-    elif animation_type == "pop":
-        return create_pop_subtitles(
-            relevant_words,
-            video_width,
-            video_height,
-            effective_template,
-            effective_font_family,
-        )
-    elif animation_type == "fade":
-        return create_fade_subtitles(
-            relevant_words,
-            video_width,
-            video_height,
-            effective_template,
-            effective_font_family,
-        )
-    else:
-        # Default static subtitles
-        return create_static_subtitles(
-            relevant_words,
-            video_width,
-            video_height,
-            effective_template,
-            effective_font_family,
-        )
-
-
-def create_static_subtitles(
-    relevant_words: List[Dict],
-    video_width: int,
-    video_height: int,
-    template: Dict,
-    font_family: str,
-) -> List[TextClip]:
-    """Create standard static subtitles (original behavior)."""
-    subtitle_clips = []
-    processor = VideoProcessor(
-        font_family, template["font_size"], template["font_color"]
-    )
-
-    calculated_font_size = get_scaled_font_size(template["font_size"], video_width)
-    position_y = template.get("position_y", 0.75)
-    max_text_width = get_subtitle_max_width(video_width)
-
-    words_per_subtitle = 3
-    for i in range(0, len(relevant_words), words_per_subtitle):
-        word_group = relevant_words[i : i + words_per_subtitle]
-        if not word_group:
-            continue
-
-        segment_start = word_group[0]["start"]
-        segment_end = word_group[-1]["end"]
-        segment_duration = segment_end - segment_start
-
-        if segment_duration < 0.1:
-            continue
-
-        text = " ".join(word["text"] for word in word_group)
-
-        try:
-            stroke_color = template.get("stroke_color", "black")
-            stroke_width = template.get("stroke_width", 1)
-
-            text_clip = (
-                TextClip(
-                    text=text,
-                    font=processor.font_path,
-                    font_size=calculated_font_size,
-                    color=template["font_color"],
-                    stroke_color=stroke_color if stroke_color else None,
-                    stroke_width=stroke_width if stroke_color else 0,
-                    method="caption",
-                    size=(max_text_width, None),
-                    text_align="center",
-                    interline=6,
-                )
-                .with_duration(segment_duration)
-                .with_start(segment_start)
-            )
-
-            text_height = text_clip.size[1] if text_clip.size else 40
-            vertical_position = get_safe_vertical_position(
-                video_height, text_height, position_y
-            )
-            text_clip = text_clip.with_position(("center", vertical_position))
-
-            subtitle_clips.append(text_clip)
-
-        except Exception as e:
-            logger.warning(f"Failed to create subtitle for '{text}': {e}")
-            continue
-
-    logger.info(f"Created {len(subtitle_clips)} static subtitle elements")
-    return subtitle_clips
-
-
-def create_karaoke_subtitles(
-    relevant_words: List[Dict],
-    video_width: int,
-    video_height: int,
-    template: Dict,
-    font_family: str,
-) -> List[TextClip]:
-    """Create karaoke-style subtitles with word-by-word highlighting."""
-    subtitle_clips = []
-    processor = VideoProcessor(
-        font_family, template["font_size"], template["font_color"]
-    )
-
-    calculated_font_size = get_scaled_font_size(template["font_size"], video_width)
-    position_y = template.get("position_y", 0.75)
-    highlight_color = template.get("highlight_color", "#FFD700")
-    normal_color = template["font_color"]
-    max_text_width = get_subtitle_max_width(video_width)
-    horizontal_padding = max(40, int(video_width * 0.06))
-
-    words_per_group = 3
-
-    def measure_word_group_width(word_group: List[Dict], font_size: int) -> List[int]:
-        widths: List[int] = []
-        for word in word_group:
-            temp_clip = TextClip(
-                text=word["text"],
-                font=processor.font_path,
-                font_size=font_size,
-                color=normal_color,
-                stroke_color=template.get("stroke_color", "black"),
-                stroke_width=template.get("stroke_width", 1),
-                method="label",
-            )
-            widths.append(temp_clip.size[0] if temp_clip.size else 50)
-            temp_clip.close()
-        return widths
-
-    for group_idx in range(0, len(relevant_words), words_per_group):
-        word_group = relevant_words[group_idx : group_idx + words_per_group]
-        if not word_group:
-            continue
-
-        group_start = word_group[0]["start"]
-        group_end = word_group[-1]["end"]
-
-        # For each word in the group, create a highlighted version
-        for word_idx, current_word in enumerate(word_group):
-            word_start = current_word["start"]
-            word_end = current_word["end"]
-            word_duration = word_end - word_start
-
-            if word_duration < 0.05:
-                continue
-
-            try:
-                # Build the text with the current word highlighted
-                # We create individual text clips for each word and composite them
-                word_clips_for_composite = []
-                font_size_for_group = calculated_font_size
-                word_widths = measure_word_group_width(word_group, font_size_for_group)
-                space_width = font_size_for_group * 0.28
-                total_width = sum(word_widths) + space_width * (len(word_group) - 1)
-
-                if total_width > max_text_width and total_width > 0:
-                    shrink_ratio = max_text_width / total_width
-                    font_size_for_group = max(
-                        20, int(font_size_for_group * shrink_ratio)
-                    )
-                    word_widths = measure_word_group_width(
-                        word_group, font_size_for_group
-                    )
-                    space_width = font_size_for_group * 0.28
-                    total_width = sum(word_widths) + space_width * (len(word_group) - 1)
-
-                # Second pass: create positioned clips
-                current_x = max(horizontal_padding, (video_width - total_width) / 2)
-                text_height = 40
-
-                for w_idx, word in enumerate(word_group):
-                    is_current = w_idx == word_idx
-                    color = highlight_color if is_current else normal_color
-                    # Scale up current word slightly for pop effect
-                    size_multiplier = 1.1 if is_current else 1.0
-
-                    word_clip = (
-                        TextClip(
-                            text=word["text"],
-                            font=processor.font_path,
-                            font_size=int(font_size_for_group * size_multiplier),
-                            color=color,
-                            stroke_color=template.get("stroke_color", "black"),
-                            stroke_width=template.get("stroke_width", 1),
-                            method="label",
-                        )
-                        .with_duration(word_duration)
-                        .with_start(word_start)
-                    )
-
-                    text_height = max(
-                        text_height, word_clip.size[1] if word_clip.size else 40
-                    )
-                    vertical_position = get_safe_vertical_position(
-                        video_height, text_height, position_y
-                    )
-
-                    word_clip = word_clip.with_position(
-                        (int(current_x), vertical_position)
-                    )
-                    word_clips_for_composite.append(word_clip)
-
-                    current_x += word_widths[w_idx] + space_width
-
-                subtitle_clips.extend(word_clips_for_composite)
-
-            except Exception as e:
-                logger.warning(
-                    f"Failed to create karaoke subtitle for word '{current_word['text']}': {e}"
-                )
-                continue
-
-    logger.info(f"Created {len(subtitle_clips)} karaoke subtitle elements")
-    return subtitle_clips
-
-
-def create_pop_subtitles(
-    relevant_words: List[Dict],
-    video_width: int,
-    video_height: int,
-    template: Dict,
-    font_family: str,
-) -> List[TextClip]:
-    """Create pop-style subtitles where each word pops in."""
-    subtitle_clips = []
-    processor = VideoProcessor(
-        font_family, template["font_size"], template["font_color"]
-    )
-
-    calculated_font_size = get_scaled_font_size(template["font_size"], video_width)
-    position_y = template.get("position_y", 0.75)
-    max_text_width = get_subtitle_max_width(video_width)
-
-    words_per_group = 3
-
-    for group_idx in range(0, len(relevant_words), words_per_group):
-        word_group = relevant_words[group_idx : group_idx + words_per_group]
-        if not word_group:
-            continue
-
-        # Show the full group text
-        group_text = " ".join(w["text"] for w in word_group)
-        group_start = word_group[0]["start"]
-        group_end = word_group[-1]["end"]
-        group_duration = group_end - group_start
-
-        if group_duration < 0.1:
-            continue
-
-        try:
-            # Create main text clip
-            text_clip = (
-                TextClip(
-                    text=group_text,
-                    font=processor.font_path,
-                    font_size=calculated_font_size,
-                    color=template["font_color"],
-                    stroke_color=template.get("stroke_color", "black"),
-                    stroke_width=template.get("stroke_width", 2),
-                    method="caption",
-                    size=(max_text_width, None),
-                    text_align="center",
-                    interline=6,
-                )
-                .with_duration(group_duration)
-                .with_start(group_start)
-            )
-
-            text_height = text_clip.size[1] if text_clip.size else 40
-            vertical_position = get_safe_vertical_position(
-                video_height, text_height, position_y
-            )
-            text_clip = text_clip.with_position(("center", vertical_position))
-
-            subtitle_clips.append(text_clip)
-
-        except Exception as e:
-            logger.warning(f"Failed to create pop subtitle: {e}")
-            continue
-
-    logger.info(f"Created {len(subtitle_clips)} pop subtitle elements")
-    return subtitle_clips
-
-
-def create_fade_subtitles(
-    relevant_words: List[Dict],
-    video_width: int,
-    video_height: int,
-    template: Dict,
-    font_family: str,
-) -> List[TextClip]:
-    """Create fade-style subtitles with smooth transitions."""
-    subtitle_clips = []
-    processor = VideoProcessor(
-        font_family, template["font_size"], template["font_color"]
-    )
-
-    calculated_font_size = get_scaled_font_size(template["font_size"], video_width)
-    position_y = template.get("position_y", 0.75)
-    has_background = template.get("background", False)
-    background_color = template.get("background_color", "#00000080")
-    max_text_width = get_subtitle_max_width(video_width)
-
-    words_per_group = 4
-
-    for group_idx in range(0, len(relevant_words), words_per_group):
-        word_group = relevant_words[group_idx : group_idx + words_per_group]
-        if not word_group:
-            continue
-
-        group_text = " ".join(w["text"] for w in word_group)
-        group_start = word_group[0]["start"]
-        group_end = word_group[-1]["end"]
-        group_duration = group_end - group_start
-
-        if group_duration < 0.1:
-            continue
-
-        try:
-            # Create text clip
-            text_clip = TextClip(
-                text=group_text,
-                font=processor.font_path,
-                font_size=calculated_font_size,
-                color=template["font_color"],
-                stroke_color=template.get("stroke_color")
-                if template.get("stroke_color")
-                else None,
-                stroke_width=template.get("stroke_width", 0),
-                method="caption",
-                size=(max_text_width, None),
-                text_align="center",
-                interline=6,
-            )
-
-            text_height = text_clip.size[1] if text_clip.size else 40
-            text_width = text_clip.size[0] if text_clip.size else 200
-            vertical_position = get_safe_vertical_position(
-                video_height, text_height, position_y
-            )
-
-            # Add background if specified
-            if has_background and background_color:
-                padding = 10
-                # Parse background color (handle alpha)
-                bg_color_hex = (
-                    background_color[:7]
-                    if len(background_color) > 7
-                    else background_color
-                )
-
-                bg_clip = (
-                    ColorClip(
-                        size=(text_width + padding * 2, text_height + padding),
-                        color=tuple(
-                            int(bg_color_hex[i : i + 2], 16) for i in (1, 3, 5)
-                        ),
-                    )
-                    .with_duration(group_duration)
-                    .with_start(group_start)
-                )
-
-                bg_clip = bg_clip.with_position(
-                    ("center", vertical_position - padding // 2)
-                )
-
-                # Apply fade to background
-                fade_duration = min(0.2, group_duration / 4)
-                bg_clip = (
-                    bg_clip.with_effects(
-                        [CrossFadeIn(fade_duration), CrossFadeOut(fade_duration)]
-                    )
-                    if group_duration > 0.5
-                    else bg_clip
-                )
-
-                subtitle_clips.append(bg_clip)
-
-            # Apply timing and position to text
-            text_clip = text_clip.with_duration(group_duration).with_start(group_start)
-            text_clip = text_clip.with_position(("center", vertical_position))
-
-            subtitle_clips.append(text_clip)
-
-        except Exception as e:
-            logger.warning(f"Failed to create fade subtitle: {e}")
-            continue
-
-    logger.info(f"Created {len(subtitle_clips)} fade subtitle elements")
-    return subtitle_clips
+# Fraction of the source height to crop away when the user marks the original
+# video as having burned-in subtitles on the top or bottom band. Trimming this
+# strip before face-tracked crop pushes the original text fully out of frame,
+# so our new subtitles can render at their normal position over clean pixels.
+ORIGINAL_SUBTITLE_BAND_FRACTION = 0.15
 
 
 def create_optimized_clip(
@@ -1161,18 +90,36 @@ def create_optimized_clip(
     font_color: str = "#FFFFFF",
     caption_template: str = "default",
     output_format: str = "vertical",
+    stroke_color_override: Optional[str] = None,
+    stroke_width_override: Optional[int] = None,
+    bold: bool = False,
+    italic: bool = False,
+    underline: bool = False,
+    avoid_original_subtitle: str = "none",
 ) -> bool:
-    """Create clip with optional subtitles. output_format: 'vertical' (9:16) or 'original' (keep source size)."""
+    """Create clip with optional subtitles.
+
+    output_format:
+      - 'vertical' (default): face-centered 9:16 crop
+      - 'fit': letterbox-blur 9:16 (full source centered on blurred background)
+      - 'original': keep source aspect ratio, no crop/resize
+      - 'capcut': same as 'original' but paired with the CapCut export path —
+        subtitles and B-roll are never burned in so the user can edit them as
+        real tracks after opening the draft in CapCut.
+    """
     try:
         duration = end_time - start_time
         if duration <= 0:
             logger.error(f"Invalid clip duration: {duration:.1f}s")
             return False
 
-        keep_original = output_format == "original"
+        # CapCut edit mode keeps the clip as-is for maximum edit-ability in
+        # CapCut; the rendering path is identical to 'original'.
+        keep_original = output_format in ("original", "capcut")
+        fit_mode = output_format == "fit"
         logger.info(
             f"Creating clip: {start_time:.1f}s - {end_time:.1f}s ({duration:.1f}s) "
-            f"subtitles={add_subtitles} template '{caption_template}' format={'original' if keep_original else 'vertical'}"
+            f"subtitles={add_subtitles} template '{caption_template}' format={output_format}"
         )
 
         # Fast path: no subtitles + original = ffmpeg stream copy (no re-encoding)
@@ -1199,7 +146,6 @@ def create_optimized_clip(
             logger.info(f"Successfully created clip (stream copy): {output_path}")
             return True
 
-        # Load and process video
         video = VideoFileClip(str(video_path))
 
         if start_time >= video.duration:
@@ -1209,29 +155,59 @@ def create_optimized_clip(
             video.close()
             return False
 
+        # Crop away the top/bottom strip that holds the original burned-in
+        # subtitles so they fall outside the frame before the 9:16 crop runs.
+        if avoid_original_subtitle in ("bottom", "top"):
+            orig_h = video.h
+            band_px = int(orig_h * ORIGINAL_SUBTITLE_BAND_FRACTION)
+            if band_px > 0 and band_px < orig_h:
+                if avoid_original_subtitle == "bottom":
+                    video = video.cropped(y1=0, y2=orig_h - band_px)
+                else:
+                    video = video.cropped(y1=band_px, y2=orig_h)
+                logger.info(
+                    f"Cropped {avoid_original_subtitle} band ({band_px}px / "
+                    f"{ORIGINAL_SUBTITLE_BAND_FRACTION * 100:.0f}%) to hide "
+                    f"original subtitles; new size {video.w}x{video.h}"
+                )
+
         end_time = min(end_time, video.duration)
         clip = video.subclipped(start_time, end_time)
 
         if keep_original:
-            # No face detection, no crop, no resize - use trimmed clip as-is
             processed_clip = clip
             target_width = round_to_even(processed_clip.w)
             target_height = round_to_even(processed_clip.h)
             if (target_width, target_height) != (processed_clip.w, processed_clip.h):
                 processed_clip = processed_clip.resized((target_width, target_height))
             cropped_clip = None
+        elif fit_mode:
+            # Fit-with-blur mode: keep the full source centered, fill the
+            # surrounding 9:16 area with a blurred zoomed copy. Face is never
+            # clipped. Isolated branch — remove to disable without touching
+            # the legacy crop path below.
+            from .fit_renderer import apply_fit_with_blur
+            composite, target_width, target_height = apply_fit_with_blur(
+                clip, target_ratio=9 / 16
+            )
+            processed_clip = composite
+            cropped_clip = None
         else:
-            # Vertical 9:16: face-centered crop, preserve native resolution
             x_offset, y_offset, new_width, new_height = detect_optimal_crop_region(
                 video, start_time, end_time, target_ratio=9 / 16
             )
             cropped_clip = clip.cropped(
-                x1=x_offset, y1=y_offset, x2=x_offset + new_width, y2=y_offset + new_height
+                x1=x_offset,
+                y1=y_offset,
+                x2=x_offset + new_width,
+                y2=y_offset + new_height,
             )
-            target_width, target_height = round_to_even(new_width), round_to_even(new_height)
+            target_width, target_height = (
+                round_to_even(new_width),
+                round_to_even(new_height),
+            )
             processed_clip = cropped_clip
 
-        # Add AssemblyAI subtitles with template support
         final_clips = [processed_clip]
 
         if add_subtitles:
@@ -1245,10 +221,15 @@ def create_optimized_clip(
                 font_size,
                 font_color,
                 caption_template,
+                stroke_color_override=stroke_color_override,
+                stroke_width_override=stroke_width_override,
+                bold=bold,
+                italic=italic,
+                underline=underline,
+                avoid_original_subtitle=avoid_original_subtitle,
             )
             final_clips.extend(subtitle_clips)
 
-        # Compose and encode
         final_clip = (
             CompositeVideoClip(final_clips) if len(final_clips) > 1 else processed_clip
         )
@@ -1266,7 +247,6 @@ def create_optimized_clip(
             **encoding_settings,
         )
 
-        # Cleanup
         if final_clip is not processed_clip:
             final_clip.close()
         if processed_clip is not cropped_clip:
@@ -1305,7 +285,6 @@ def create_clips_from_segments(
 
     for i, segment in enumerate(segments):
         try:
-            # Debug log the segment data
             logger.info(
                 f"Processing segment {i + 1}: start='{segment.get('start_time')}', end='{segment.get('end_time')}'"
             )
@@ -1351,7 +330,6 @@ def create_clips_from_segments(
                     "text": segment["text"],
                     "relevance_score": segment["relevance_score"],
                     "reasoning": segment["reasoning"],
-                    # Include virality data if available
                     "virality_score": segment.get("virality_score", 0),
                     "hook_score": segment.get("hook_score", 0),
                     "engagement_score": segment.get("engagement_score", 0),
@@ -1369,118 +347,6 @@ def create_clips_from_segments(
 
     logger.info(f"Successfully created {len(clips_info)}/{len(segments)} clips")
     return clips_info
-
-
-def get_available_transitions() -> List[str]:
-    """Get list of available transition video files."""
-    transitions_dir = Path(__file__).parent.parent / "transitions"
-    if not transitions_dir.exists():
-        logger.warning("Transitions directory not found")
-        return []
-
-    transition_files = []
-    for file_path in transitions_dir.glob("*.mp4"):
-        transition_files.append(str(file_path))
-
-    logger.info(f"Found {len(transition_files)} transition files")
-    return transition_files
-
-
-def apply_transition_effect(
-    clip1_path: Path, clip2_path: Path, transition_path: Path, output_path: Path
-) -> bool:
-    """Apply transition effect between two clips using a transition video."""
-    clip1 = None
-    clip2 = None
-    transition = None
-    clip1_tail = None
-    clip2_intro = None
-    clip2_remainder = None
-    intro_segment = None
-    final_clip = None
-
-    try:
-        from moviepy import VideoFileClip, CompositeVideoClip, concatenate_videoclips
-
-        # Load clips
-        clip1 = VideoFileClip(str(clip1_path))
-        clip2 = VideoFileClip(str(clip2_path))
-        transition = VideoFileClip(str(transition_path))
-
-        # Keep the transition window within both clips so the output still matches
-        # the current clip's duration and metadata.
-        transition_duration = min(1.5, transition.duration, clip1.duration, clip2.duration)
-        if transition_duration <= 0:
-            logger.warning("Transition duration is zero, skipping transition effect")
-            return False
-
-        transition = transition.subclipped(0, transition_duration)
-
-        # Resize transition to match clip dimensions
-        clip_size = clip2.size
-        transition = transition.resized(clip_size)
-
-        # Build a transition intro from the previous clip tail over the first
-        # part of the current clip so the exported file keeps clip2's duration.
-        clip1_tail_start = max(0, clip1.duration - transition_duration)
-        clip1_tail = clip1.subclipped(clip1_tail_start, clip1.duration).with_effects(
-            [FadeOut(transition_duration)]
-        )
-        clip2_intro = clip2.subclipped(0, transition_duration).with_effects(
-            [FadeIn(transition_duration)]
-        )
-
-        intro_segment = CompositeVideoClip(
-            [clip1_tail, clip2_intro, transition], size=clip_size
-        ).with_duration(transition_duration)
-        if clip2_intro.audio is not None:
-            intro_segment = intro_segment.with_audio(clip2_intro.audio)
-
-        final_segments = [intro_segment]
-        if clip2.duration > transition_duration:
-            clip2_remainder = clip2.subclipped(transition_duration, clip2.duration)
-            final_segments.append(clip2_remainder)
-
-        final_clip = (
-            concatenate_videoclips(final_segments, method="compose")
-            if len(final_segments) > 1
-            else intro_segment
-        )
-
-        # Write output
-        processor = VideoProcessor()
-        encoding_settings = processor.get_optimal_encoding_settings("high")
-
-        final_clip.write_videofile(
-            str(output_path),
-            temp_audiofile="temp-audio.m4a",
-            remove_temp=True,
-            logger=None,
-            **encoding_settings,
-        )
-
-        logger.info(f"Applied transition effect: {output_path}")
-        return True
-
-    except Exception as e:
-        logger.error(f"Error applying transition effect: {e}")
-        return False
-    finally:
-        for clip in (
-            final_clip,
-            intro_segment,
-            clip2_remainder,
-            clip2_intro,
-            clip1_tail,
-            transition,
-            clip2,
-            clip1,
-        ):
-            if clip is not None:
-                try:
-                    clip.close()
-                except Exception:
-                    pass
 
 
 def create_clips_with_transitions(
@@ -1517,243 +383,6 @@ def create_clips_with_transitions(
     )
 
 
-# Backward compatibility functions
 def get_video_transcript_with_assemblyai(path: Path) -> str:
     """Backward compatibility wrapper."""
     return get_video_transcript(path)
-
-
-def create_9_16_clip(
-    video_path: Path,
-    start_time: float,
-    end_time: float,
-    output_path: Path,
-    subtitle_text: str = "",
-) -> bool:
-    """Backward compatibility wrapper."""
-    return create_optimized_clip(
-        video_path, start_time, end_time, output_path, add_subtitles=bool(subtitle_text)
-    )
-
-
-# B-Roll compositing functions
-
-
-def insert_broll_into_clip(
-    main_clip_path: Path,
-    broll_path: Path,
-    insert_time: float,
-    broll_duration: float,
-    output_path: Path,
-    transition_duration: float = 0.3,
-) -> bool:
-    """
-    Insert B-roll footage into a clip at a specified timestamp.
-
-    Args:
-        main_clip_path: Path to the main video clip
-        broll_path: Path to the B-roll video
-        insert_time: When to insert B-roll (seconds from clip start)
-        broll_duration: How long to show B-roll (seconds)
-        output_path: Where to save the composited clip
-        transition_duration: Crossfade duration (seconds)
-
-    Returns:
-        True if successful
-    """
-    try:
-        from moviepy import VideoFileClip, CompositeVideoClip, concatenate_videoclips
-        from moviepy.video.fx import CrossFadeIn, CrossFadeOut
-
-        # Load clips
-        main_clip = VideoFileClip(str(main_clip_path))
-        broll_clip = VideoFileClip(str(broll_path))
-
-        # Get main clip dimensions
-        target_width, target_height = main_clip.size
-
-        # Resize B-roll to match main clip (9:16 aspect ratio)
-        broll_resized = resize_for_916(broll_clip, target_width, target_height)
-
-        # Ensure B-roll doesn't exceed requested duration
-        actual_broll_duration = min(broll_duration, broll_resized.duration)
-        broll_trimmed = broll_resized.subclipped(0, actual_broll_duration)
-
-        # Ensure insert_time is within clip bounds
-        insert_time = max(0, min(insert_time, main_clip.duration - 0.5))
-
-        # Calculate end time for B-roll
-        broll_end_time = insert_time + actual_broll_duration
-
-        # Don't let B-roll extend past the main clip
-        if broll_end_time > main_clip.duration:
-            broll_end_time = main_clip.duration
-            actual_broll_duration = broll_end_time - insert_time
-            broll_trimmed = broll_resized.subclipped(0, actual_broll_duration)
-
-        # Split main clip into three parts
-        part1 = main_clip.subclipped(0, insert_time) if insert_time > 0 else None
-        part2_audio = main_clip.subclipped(insert_time, broll_end_time).audio
-        part3 = (
-            main_clip.subclipped(broll_end_time)
-            if broll_end_time < main_clip.duration
-            else None
-        )
-
-        # Apply crossfade to B-roll
-        if transition_duration > 0:
-            broll_with_audio = broll_trimmed.with_audio(part2_audio)
-            broll_faded = broll_with_audio.with_effects(
-                [CrossFadeIn(transition_duration), CrossFadeOut(transition_duration)]
-            )
-        else:
-            broll_faded = broll_trimmed.with_audio(part2_audio)
-
-        # Concatenate parts
-        clips_to_concat = []
-        if part1:
-            clips_to_concat.append(part1)
-        clips_to_concat.append(broll_faded)
-        if part3:
-            clips_to_concat.append(part3)
-
-        if len(clips_to_concat) == 1:
-            final_clip = clips_to_concat[0]
-        else:
-            final_clip = concatenate_videoclips(clips_to_concat, method="compose")
-
-        # Write output
-        processor = VideoProcessor()
-        encoding_settings = processor.get_optimal_encoding_settings("high")
-
-        final_clip.write_videofile(
-            str(output_path),
-            temp_audiofile="temp-audio-broll.m4a",
-            remove_temp=True,
-            logger=None,
-            **encoding_settings,
-        )
-
-        # Cleanup
-        final_clip.close()
-        main_clip.close()
-        broll_clip.close()
-        broll_resized.close()
-
-        logger.info(
-            f"Inserted B-roll at {insert_time:.1f}s ({actual_broll_duration:.1f}s duration): {output_path}"
-        )
-        return True
-
-    except Exception as e:
-        logger.error(f"Error inserting B-roll: {e}")
-        return False
-
-
-def resize_for_916(
-    clip: VideoFileClip, target_width: int, target_height: int
-) -> VideoFileClip:
-    """
-    Resize a video clip to fit 9:16 aspect ratio with center crop.
-
-    Args:
-        clip: Input video clip
-        target_width: Target width
-        target_height: Target height
-
-    Returns:
-        Resized video clip
-    """
-    clip_width, clip_height = clip.size
-    target_aspect = target_width / target_height
-    clip_aspect = clip_width / clip_height
-
-    if clip_aspect > target_aspect:
-        # Clip is wider - scale to height and crop width
-        scale_factor = target_height / clip_height
-        new_width = int(clip_width * scale_factor)
-        new_height = target_height
-        resized = clip.resized((new_width, new_height))
-
-        # Center crop
-        x_offset = (new_width - target_width) // 2
-        cropped = resized.cropped(x1=x_offset, x2=x_offset + target_width)
-    else:
-        # Clip is taller - scale to width and crop height
-        scale_factor = target_width / clip_width
-        new_width = target_width
-        new_height = int(clip_height * scale_factor)
-        resized = clip.resized((new_width, new_height))
-
-        # Center crop (crop from top for portrait videos)
-        y_offset = (new_height - target_height) // 4  # Bias towards top
-        cropped = resized.cropped(y1=y_offset, y2=y_offset + target_height)
-
-    return cropped
-
-
-def apply_broll_to_clip(
-    clip_path: Path, broll_suggestions: List[Dict[str, Any]], output_path: Path
-) -> bool:
-    """
-    Apply multiple B-roll insertions to a clip.
-
-    Args:
-        clip_path: Path to the main clip
-        broll_suggestions: List of B-roll suggestions with local_path, timestamp, duration
-        output_path: Where to save the final clip
-
-    Returns:
-        True if successful
-    """
-    if not broll_suggestions:
-        logger.info("No B-roll suggestions to apply")
-        return False
-
-    try:
-        # Sort suggestions by timestamp (process from end to start to preserve timing)
-        sorted_suggestions = sorted(
-            broll_suggestions, key=lambda x: x.get("timestamp", 0), reverse=True
-        )
-
-        current_clip_path = clip_path
-        temp_paths = []
-
-        for i, suggestion in enumerate(sorted_suggestions):
-            broll_path = suggestion.get("local_path")
-            if not broll_path or not Path(broll_path).exists():
-                logger.warning(f"B-roll file not found: {broll_path}")
-                continue
-
-            timestamp = suggestion.get("timestamp", 0)
-            duration = suggestion.get("duration", 3.0)
-
-            # Create temp output for intermediate clips
-            if i < len(sorted_suggestions) - 1:
-                temp_output = output_path.parent / f"temp_broll_{i}.mp4"
-                temp_paths.append(temp_output)
-            else:
-                temp_output = output_path
-
-            success = insert_broll_into_clip(
-                current_clip_path, Path(broll_path), timestamp, duration, temp_output
-            )
-
-            if success:
-                current_clip_path = temp_output
-            else:
-                logger.warning(f"Failed to insert B-roll at {timestamp}s")
-
-        # Cleanup temp files
-        for temp_path in temp_paths:
-            if temp_path.exists() and temp_path != output_path:
-                try:
-                    temp_path.unlink()
-                except Exception:
-                    pass
-
-        return True
-
-    except Exception as e:
-        logger.error(f"Error applying B-roll to clip: {e}")
-        return False
